Add feature to support a "Max Cool On Time"
This will turn off the cool output if it has been on for too long.
Specifically being used to support a coolant circulation pump that if it runs for too long the heat from the pump warms up the cooling reservoir, and ends up raising the whole temp of the system, rather than making it cooler.
This prevents thermal runaway.